feat(plexos): plexos parser xml (#93)
Update the plexos parser to find a single xml file in the run_directory
when available. If the fmap is provided that will be used in favor of
finding the only xml file. When multiple xml files are present, this
will raise an Error, and then the fmap will be required again.