feat(models): Updating PHES to reflect latest version of PSY. (#144)
This PR adds the correct model field for `storage_capacity` for HydroPumpedStorage and also changes the way we create named tuples. It turns out that if we create classes, the serialization is easier without to much compromise. Pydantic knows how to create the dict and deserialize it allowing us to deprecate the custom serialization.